requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.6
django-admin-interface==0.30.1
django-cleanup==9.0.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-modeltranslation==0.19.16
django-oauth-toolkit==3.0.1
django-oauth2-provider==0.2.6.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.8.2
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
shortuuid==1.0.13
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
